@@ -13,12 +13,12 @@
     "gulp-cssimport": "^1.3.1",
     "gulp-jade": "^1.0.0",
     "gulp-stylus": "^2.0.0",
-    "gulp-typescript": "~2.6.0",
+    "gulp-typescript": "~2.8.3",
     "gulp-util": "^3.0.3",
     "nib": "^1.1.0",
-    "typescript": "1.5.0-alpha",
+    "typescript": "1.6.2",
     "vinyl-source-stream": "^1.0.0",
-    "watchify": "^2.3.0"
+    "watchify": "^3.4.0"
   },
   "dependencies": {
     "async": "^0.9.0",
