@@ -1,9 +1,9 @@
 /// <reference path="async/async.d.ts" />
+/// <reference path="express/express.d.ts" />
 /// <reference path="lodash/lodash.d.ts" />
 /// <reference path="node/node.d.ts" />
-/// <reference path="express/express.d.ts" />
+/// <reference path="recursive-readdir/recursive-readdir.d.ts" />
 /// <reference path="rimraf/rimraf.d.ts" />
 /// <reference path="socket.io/socket.io.d.ts" />
 /// <reference path="tv4/tv4.d.ts" />
 /// <reference path="yargs/yargs.d.ts" />
-/// <reference path="recursive-readdir/recursive-readdir.d.ts" />
